refactor: extract message.py from participant.py
Move add_message, _find_insertion_index, and _find_closest_participant
to their own module. Mirrors the frontend split (sequence-message.js
vs sequence-operations.js).
