+++ .kiro/steering/structure.md
@@ -14,8 +14,9 @@
 │   │   └── parse_changelog.py # CHANGELOG.md parser for version history
 │   ├── sequence/           # Sequence diagram package
 │   │   ├── routes.py       # Sequence routes (/addParticipant, /addMessage, etc.)
-│   │   ├── classes.py      # Diagram, Participant data classes
-│   │   └── participant.py  # Participant logic
+│   │   ├── classes.py      # Diagram, Participant, Message data classes
+│   │   ├── participant.py  # Participant logic (add, rename, delete, positions)
+│   │   └── message.py      # Message logic (add message, y-based insertion)
 │   ├── activity/           # Activity diagram package
 │   │   ├── routes.py       # All activity routes (~64 endpoints)
 │   │   ├── classes.py      # RectElement, PolyElement, Ellipse, SvgChunk, TextElement
